Fixed MAPE error, added pred plot
Using MAPE was causing a problem if a data point actual value was 0, which is true for at least one value with 0-1 scaling, the percent error would blow up and cause erroneous values in the metric.  Changed to MAE for now to avoid this.  Fixed a few other things, most important was to make test data always at the end of the time series, code is now ready to experiment to see if I can get decent accuracy.